revise the way data is handled
on the js side, store all items in a flat data store so that if e.g. an
order is refreshed, other cards on the page that are showing this order
are also refreshed. (this often does not behave as expected, because
the fields for table views of a model may differ from those in a full
card view, so refreshing the latter doesn’t affect the former even
though an update did take place.

on the ruby side, have #act return json with top-level ‘result’ and
‘changed’ keys: result is the return value from form.execute and
changed is a list of objects that were changed. #update just returns
the latter. this is preparation for
https://app.asana.com/0/9610797295787/9430395475895

Deckhand::Form implementers are now required to implement
#changed_objects, which is a list of instances that will be used in the
JSON response above.