Deckhand::Form.multiple
allows the creation of forms that let you add and remove inputs in
groups, viz:

class Animal < Deckhand::Form
  multiple :noises do
    input :name
    input :volume, type: Integer
  end
end

which produces a UI that lets you edit the name and volume for each
noise in a list of noises, and add or remove them.

upon submission, the multiple field’s accessor will be set to a list of
hashes (actually ActiveSupport::HashWithIndifferentAccess), with values
cast to correct types if specified.